Refactored snooze button into separate react component.  Added quick snooze buttons.
