+++ client/src/components/reminders/Reminders.jsx
@@ -1,12 +1,11 @@
+import {ActionCreators} from '../../actions';
 import Badge from '@material-ui/core/Badge';
 import BaseTodoList from '../BaseTodoList';
-import Button from '@material-ui/core/Button';
 import Checkbox from '@material-ui/core/Checkbox';
 import Divider from '@material-ui/core/Divider';
 import EditIcon from '@material-ui/icons/Edit';
 import EditTodoDialog from '../edit-dialog';
 import LoopIcon from '@material-ui/icons/Loop';
-import SnoozeIcon from '@material-ui/icons/Snooze';
 import IconButton from '@material-ui/core/IconButton';
 import List from '@material-ui/core/List';
 import ListItem from '@material-ui/core/ListItem';
@@ -14,49 +13,14 @@ import ListItemText from '@material-ui/core/ListItemText';
 import Paper from '@material-ui/core/Paper';
 import Tooltip from '@material-ui/core/Tooltip';
 import Typography from '@material-ui/core/Typography';
-import Popover from '@material-ui/core/Popover';
 import PropTypes from 'prop-types';
 import React from 'react';
+import SnoozeButton from './SnoozeButton';
 import Sugar from 'sugar-date';
 import Toolbar from '../toolbar/Toolbar';
 import {withStyles} from '@material-ui/core/styles';
 import {bindActionCreators, compose} from 'redux';
-import {ActionCreators} from '../../actions';
 import {connect} from 'react-redux';
-
-
-const snoozeOptions = {
-  '5m': 'in 5 minutes',
-  '15m': 'in 15 minutes',
-  '30m': 'in 30 minutes',
-  '45m': 'in 45 minutes',
-  '1h': 'in 1 hour',
-  '2h': 'in 2 hours',
-  '3h': 'in 3 hours',
-  '4h': 'in 4 hours',
-  '8h': 'in 8 hours',
-  '12h': 'in 12 hours',
-  '20h': 'in 20 hours',
-  '1d': 'in 1 day',
-  '2d': 'in 2 days',
-  '3d': 'in 3 days',
-  '1w': 'in 1 week',
-  '2w': 'in 2 weeks',
-  '1mo': 'in 1 month',
-  '3mo': 'in 3 months',
-  '6mo': 'in 6 months',
-  '1y': 'in 1 year',
-  'morrow morn': 'tomorrow at 9am',
-  'morrow noon': 'tomorrow at noon',
-  'morrow eve': 'tomorrow at 5pm',
-  'Su': 'next Sunday at noon',
-  'M': 'next Monday at noon',
-  'Tu': 'next Tuesday at noon',
-  'W': 'next Wednesday at noon',
-  'Th': 'next Thursday at noon',
-  'F': 'next Friday at noon',
-  'Sa': 'next Saturday at noon',
-};
 
 const styles = theme => ({
   root: {
@@ -81,11 +45,6 @@ const styles = theme => ({
     right: -8,
     width: 15,
     height: 15,
-  },
-  popoverTitle: {
-    padding: 10,
-    backgroundColor: theme.palette.primary.main,
-    color: theme.palette.common.white,
   },
 });
 
@@ -115,33 +74,18 @@ function getSorting(order, orderBy) {
 
 class Reminders extends BaseTodoList {
   state = {
-    anchorEl: null,
     orderBy: 'nextReminder',
     order: 'desc',
     searchTxt: null,
     editing: false,
     remindersOnly: true,
-};
+  };
 
-  handleSnooze = async e => {
-    this.hideSnoozePopover();
-    const next = Sugar.Date.create(e.currentTarget.value);
+  handleSnooze = async (id, value) => {
+    const next = Sugar.Date.create(value);
     await this.props.actions.updateTodo(
-      this.state.selected, {nextReminder: next.toISOString()});
+      id, {nextReminder: next.toISOString()});
   };
-
-  showSnoozePopover = (target, id, headline) => {
-    this.setState({
-      anchorEl: target,
-      selected: id,
-      selectedHeadline: headline,
-    });
-  };
-
-  hideSnoozePopover = () => {
-    this.setState({anchorEl: null});
-  };
-
 
   handleSnoozeAll = () => {
     this.props.actions.snoozeAll(300);
@@ -154,8 +98,7 @@ class Reminders extends BaseTodoList {
     }
     todos = todos.filter(t => !t.complete && t.nextReminder &&
       new Date(t.nextReminder) <= new Date());
-    const {filtered, order, orderBy, anchorEl, editing} = this.state;
-    const open = Boolean(anchorEl);
+    const {filtered, order, orderBy, editing} = this.state;
 
     const abbreviatedRepeat =repeat => {
       let abbrev = repeat.charAt(0).toUpperCase();
@@ -200,14 +143,10 @@ class Reminders extends BaseTodoList {
                       }
                     />
                     <IconButton aria-label="Edit" className={classes.editControl}
-                                onClick={e => this.setState({editing: n})}>
+                                onClick={() => this.setState({editing: n})}>
                       <EditIcon/>
                     </IconButton>
-                    <IconButton aria-label="Snooze" className={classes.editControl}
-                                onClick={e => this.showSnoozePopover(e.currentTarget, n.id, n.headline)}
-                    >
-                      <SnoozeIcon/>
-                    </IconButton>
+                    <SnoozeButton reminder={n} onClick={this.handleSnooze}/>
                   </ListItem>
                   <li>
                     <Divider variant="inset"/>
@@ -217,29 +156,6 @@ class Reminders extends BaseTodoList {
             })
           }
         </List>
-        <Popover
-          id="simple-popper"
-          open={open}
-          anchorEl={anchorEl}
-          onClose={this.hideSnoozePopover}
-          anchorOrigin={{
-            vertical: 'bottom',
-            horizontal: 'center',
-          }}
-          transformOrigin={{
-            vertical: 'top',
-            horizontal: 'center',
-          }}
-        >
-          <Typography className={classes.popoverTitle} variant="subheading" component="h5">
-            {`Snooze: ${this.state.selectedHeadline}`}
-          </Typography>
-          <Divider/>
-          {Object.keys(snoozeOptions).map(opt => (
-            <Button key={opt} value={snoozeOptions[opt]}
-                    onClick={this.handleSnooze}>{opt}</Button>
-          ))}
-        </Popover>
         <EditTodoDialog todo={editing} onSave={this.handleSave}
                         onClose={this.handleCloseEditDialog}
                         onDelete={this.handleDelete}
